fix std::bad_variant_access for uint32_t EID comparison
invalidEid holds uint8_t(0xFF), so std::get<uint32_t>(invalidEid)
throws std::bad_variant_access, causing getEid() to silently return
nullopt for all uint32_t EIDs. Use std::get<uint8_t> with a cast to
match the active alternative.

Tested: SPDM sanity across all the endpoints and the results
are as expected

Fixes nvbug https://nvbugspro.nvidia.com/bug/6402067

+++ spdmd/mctp_endpoint_discovery.cpp
@@ -717,7 +717,8 @@ MctpDiscovery::EidType
             }
         }
         else if (std::holds_alternative<uint32_t>(*eid) &&
-                 std::get<uint32_t>(*eid) < std::get<uint32_t>(invalidEid))
+                 std::get<uint32_t>(*eid) <
+                     static_cast<uint32_t>(std::get<uint8_t>(invalidEid)))
         {
             try
             {
